basic functions completed

+++ app/src/main/java/com/myself/mychallengueapp/data/model/habitLogs.kt
@@ -7,6 +7,6 @@ import java.time.LocalDate
 @Entity(tableName = "habit_logs")
 data class habitLogs(
     @PrimaryKey(autoGenerate = true) val logId: Long = 0,
-    val habitId: Long,
+    val habitId: Int,
     val date: LocalDate
 )
